Cleaned up main package by moving top-level classes to their correct locations
