+++ app/src/main/java/org/pattonvillecs/pattonvilleapp/di/ActivityBuilderModule.java
@@ -17,8 +17,8 @@
 
 package org.pattonvillecs.pattonvilleapp.di;
 
-import org.pattonvillecs.pattonvilleapp.MainActivity;
 import org.pattonvillecs.pattonvilleapp.di.scopes.PerActivity;
+import org.pattonvillecs.pattonvilleapp.view.ui.MainActivity;
 import org.pattonvillecs.pattonvilleapp.view.ui.about.detail.AboutDetailActivity;
 import org.pattonvillecs.pattonvilleapp.view.ui.calendar.details.CalendarEventDetailsActivity;
 import org.pattonvillecs.pattonvilleapp.view.ui.directory.detail.all.AllDataSourcesDirectoryDetailActivity;
