+++ app/src/main/java/org/pattonvillecs/pattonvilleapp/service/model/calendar/event/PinnableCalendarEvent.kt
@@ -24,7 +24,7 @@ import android.arch.persistence.room.Relation
 import android.os.Parcel
 import android.os.Parcelable
 import com.google.errorprone.annotations.Immutable
-import org.pattonvillecs.pattonvilleapp.DataSource
+import org.pattonvillecs.pattonvilleapp.service.model.DataSource
 import org.pattonvillecs.pattonvilleapp.service.model.calendar.DataSourceMarker
 
 /**
