+++ app/src/main/java/org/pattonvillecs/pattonvilleapp/service/repository/directory/DirectoryDao.java
@@ -26,7 +26,7 @@ import android.arch.persistence.room.Transaction;
 import android.arch.persistence.room.Update;
 import android.support.annotation.NonNull;
 
-import org.pattonvillecs.pattonvilleapp.DataSource;
+import org.pattonvillecs.pattonvilleapp.service.model.DataSource;
 import org.pattonvillecs.pattonvilleapp.service.model.directory.Faculty;
 
 import java.util.List;
