+++ app/src/main/java/org/pattonvillecs/pattonvilleapp/service/repository/news/NewsDao.java
@@ -26,7 +26,7 @@ import android.arch.persistence.room.Transaction;
 import android.arch.persistence.room.Update;
 import android.support.annotation.NonNull;
 
-import org.pattonvillecs.pattonvilleapp.DataSource;
+import org.pattonvillecs.pattonvilleapp.service.model.DataSource;
 import org.pattonvillecs.pattonvilleapp.service.model.news.ArticleSummary;
 
 import java.util.List;
