+++ app/src/main/java/org/pattonvillecs/pattonvilleapp/view/ui/calendar/PinnableCalendarEventItem.kt
@@ -45,8 +45,8 @@ import me.xdrop.fuzzywuzzy.FuzzySearch
 import org.jetbrains.anko.coroutines.experimental.bg
 import org.jetbrains.anko.find
 import org.jetbrains.anko.sdk25.coroutines.onClick
-import org.pattonvillecs.pattonvilleapp.DataSource
 import org.pattonvillecs.pattonvilleapp.R
+import org.pattonvillecs.pattonvilleapp.service.model.DataSource
 import org.pattonvillecs.pattonvilleapp.service.model.calendar.event.ICalendarEvent
 import org.pattonvillecs.pattonvilleapp.service.model.calendar.event.PinnableCalendarEvent
 import org.pattonvillecs.pattonvilleapp.view.ui.calendar.PinnableCalendarEventItem.PinnableCalendarEventItemViewHolder
