+++ app/src/main/java/org/pattonvillecs/pattonvilleapp/view/ui/directory/detail/all/DataSourceHeader.kt
@@ -23,8 +23,8 @@ import eu.davidea.flexibleadapter.FlexibleAdapter
 import eu.davidea.flexibleadapter.items.AbstractHeaderItem
 import eu.davidea.flexibleadapter.items.IFlexible
 import eu.davidea.viewholders.FlexibleViewHolder
-import org.pattonvillecs.pattonvilleapp.DataSource
 import org.pattonvillecs.pattonvilleapp.R
+import org.pattonvillecs.pattonvilleapp.service.model.DataSource
 import org.pattonvillecs.pattonvilleapp.view.ui.directory.detail.IFacultyItem
 import org.pattonvillecs.pattonvilleapp.view.ui.directory.detail.all.DataSourceHeader.DataSourceHeaderViewHolder
