+++ app/src/main/java/org/pattonvillecs/pattonvilleapp/view/ui/directory/detail/single/DataSourceSummaryItem.kt
@@ -32,8 +32,8 @@ import eu.davidea.viewholders.FlexibleViewHolder
 import org.jetbrains.anko.browse
 import org.jetbrains.anko.find
 import org.jetbrains.anko.sdk25.coroutines.onClick
-import org.pattonvillecs.pattonvilleapp.DataSource
 import org.pattonvillecs.pattonvilleapp.R
+import org.pattonvillecs.pattonvilleapp.service.model.DataSource
 import org.pattonvillecs.pattonvilleapp.view.ui.calendar.details.CalendarEventDetailsActivity
 import org.pattonvillecs.pattonvilleapp.view.ui.directory.detail.IFacultyItem
 import org.pattonvillecs.pattonvilleapp.view.ui.directory.detail.single.DataSourceSummaryItem.DataSourceSummaryItemViewHolder
