+++ app/src/main/java/org/pattonvillecs/pattonvilleapp/viewmodel/calendar/month/CalendarMonthFragmentViewModel.kt
@@ -22,8 +22,8 @@ import android.arch.lifecycle.AndroidViewModel
 import android.arch.lifecycle.LiveData
 import android.arch.lifecycle.MutableLiveData
 import com.google.common.collect.Multiset
-import org.pattonvillecs.pattonvilleapp.DataSource
 import org.pattonvillecs.pattonvilleapp.preferences.PreferenceUtils
+import org.pattonvillecs.pattonvilleapp.service.model.DataSource
 import org.pattonvillecs.pattonvilleapp.service.repository.calendar.CalendarRepository
 import org.pattonvillecs.pattonvilleapp.view.ui.calendar.PinnableCalendarEventItem
 import org.pattonvillecs.pattonvilleapp.view.ui.calendar.map
