+++ app/src/main/java/org/pattonvillecs/pattonvilleapp/viewmodel/directory/detail/single/SingleDataSourceDirectoryDetailActivityViewModel.kt
@@ -21,7 +21,7 @@ import android.app.Application
 import android.arch.lifecycle.AndroidViewModel
 import android.arch.lifecycle.LiveData
 import android.arch.lifecycle.MutableLiveData
-import org.pattonvillecs.pattonvilleapp.DataSource
+import org.pattonvillecs.pattonvilleapp.service.model.DataSource
 import org.pattonvillecs.pattonvilleapp.service.repository.directory.DirectoryRepository
 import org.pattonvillecs.pattonvilleapp.view.ui.calendar.map
 import org.pattonvillecs.pattonvilleapp.view.ui.calendar.switchMap
