+++ app/src/main/res/layout-v21/app_bar_main.xml
@@ -6,7 +6,7 @@
     android:layout_height="match_parent"
     android:fitsSystemWindows="true"
     android:orientation="vertical"
-    tools:context="org.pattonvillecs.pattonvilleapp.MainActivity">
+    tools:context="org.pattonvillecs.pattonvilleapp.view.ui.MainActivity">
 
     <android.support.design.widget.AppBarLayout
         android:layout_width="match_parent"
