+++ app/src/main/res/layout/content_default.xml
@@ -6,7 +6,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     app:layout_behavior="@string/appbar_scrolling_view_behavior"
-    tools:context="org.pattonvillecs.pattonvilleapp.MainActivity"
+    tools:context="org.pattonvillecs.pattonvilleapp.view.ui.MainActivity"
     tools:showIn="@layout/app_bar_main">
 
 </RelativeLayout>
